+++ cloudapi-20160714/core/Models/DeleteLogConfigRequest.cs
@@ -9,6 +9,9 @@ using Tea;
 namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 {
     public class DeleteLogConfigRequest : TeaModel {
+        /// <summary>
+        /// The log type. Valid values: **log** and **survey**.
+        /// </summary>
         [NameInMap("LogType")]
         [Validation(Required=false)]
         public string LogType { get; set; }