+++ cloudapi-20160714/core/Models/DescribeApiHistoryRequest.cs
@@ -39,7 +39,7 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
         /// 
         /// *   **RELEASE**
         /// *   **TEST**: the test environment
-        /// *   PRE: the pre-release environment
+        /// *   **PRE**: the pre-release environment
         /// </summary>
         [NameInMap("StageName")]
         [Validation(Required=false)]