+++ cloudapi-20160714/core/Models/DescribeApisWithStageNameIntegratedByAppRequest.cs
@@ -9,34 +9,58 @@ using Tea;
 namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 {
     public class DescribeApisWithStageNameIntegratedByAppRequest : TeaModel {
+        /// <summary>
+        /// The API name.
+        /// </summary>
         [NameInMap("ApiName")]
         [Validation(Required=false)]
         public string ApiName { get; set; }
 
+        /// <summary>
+        /// The API ID.
+        /// </summary>
         [NameInMap("ApiUid")]
         [Validation(Required=false)]
         public string ApiUid { get; set; }
 
+        /// <summary>
+        /// The application ID.
+        /// </summary>
         [NameInMap("AppId")]
         [Validation(Required=false)]
         public long? AppId { get; set; }
 
+        /// <summary>
+        /// The API description. The description can be up to 200 characters in length.
+        /// </summary>
         [NameInMap("Description")]
         [Validation(Required=false)]
         public string Description { get; set; }
 
+        /// <summary>
+        /// The request HTTP method of the API.
+        /// </summary>
         [NameInMap("Method")]
         [Validation(Required=false)]
         public string Method { get; set; }
 
+        /// <summary>
+        /// The page number of the page to return.
+        /// </summary>
         [NameInMap("PageNumber")]
         [Validation(Required=false)]
         public int? PageNumber { get; set; }
 
+        /// <summary>
+        /// The number of entries per page.
+        /// </summary>
         [NameInMap("PageSize")]
         [Validation(Required=false)]
         public int? PageSize { get; set; }
 
+        /// <summary>
+        /// The request path of the API.
+        /// </summary>
         [NameInMap("Path")]
         [Validation(Required=false)]
         public string Path { get; set; }