+++ cloudapi-20160714/core/Models/DescribeDatasetListRequest.cs
@@ -9,14 +9,23 @@ using Tea;
 namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 {
     public class DescribeDatasetListRequest : TeaModel {
+        /// <summary>
+        /// The IDs of the datasets.
+        /// </summary>
         [NameInMap("DatasetIds")]
         [Validation(Required=false)]
         public string DatasetIds { get; set; }
 
+        /// <summary>
+        /// The number of the page to return. Pages start from page 1. Default value: 1.
+        /// </summary>
         [NameInMap("PageNumber")]
         [Validation(Required=false)]
         public int? PageNumber { get; set; }
 
+        /// <summary>
+        /// The number of entries returned per page.
+        /// </summary>
         [NameInMap("PageSize")]
         [Validation(Required=false)]
         public int? PageSize { get; set; }
@@ -25,14 +34,23 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
         [Validation(Required=false)]
         public string SecurityToken { get; set; }
 
+        /// <summary>
+        /// The tag of objects that match the rule. You can specify multiple tags.
+        /// </summary>
         [NameInMap("Tag")]
         [Validation(Required=false)]
         public List<DescribeDatasetListRequestTag> Tag { get; set; }
         public class DescribeDatasetListRequestTag : TeaModel {
+            /// <summary>
+            /// The tag key.
+            /// </summary>
             [NameInMap("Key")]
             [Validation(Required=false)]
             public string Key { get; set; }
 
+            /// <summary>
+            /// The tag value.
+            /// </summary>
             [NameInMap("Value")]
             [Validation(Required=false)]
             public string Value { get; set; }