+++ cloudapi-20160714/core/Models/DescribeDatasetListResponseBody.cs
@@ -9,6 +9,9 @@ using Tea;
 namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 {
     public class DescribeDatasetListResponseBody : TeaModel {
+        /// <summary>
+        /// The returned dataset information. It is an array consisting of datasetinfo.
+        /// </summary>
         [NameInMap("DatasetInfoList")]
         [Validation(Required=false)]
         public List<DescribeDatasetListResponseBodyDatasetInfoList> DatasetInfoList { get; set; }
@@ -21,24 +24,24 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
             public string CreatedTime { get; set; }
 
             /// <summary>
-            /// The ID of the dataset.
+            /// The dataset ID.
             /// </summary>
             [NameInMap("DatasetId")]
             [Validation(Required=false)]
             public string DatasetId { get; set; }
 
             /// <summary>
-            /// The name of the dataset.
+            /// The dataset name.
             /// </summary>
             [NameInMap("DatasetName")]
             [Validation(Required=false)]
             public string DatasetName { get; set; }
 
             /// <summary>
-            /// The type of the dataset. Valid values:
+            /// The dataset type. Valid values:
             /// 
-            /// *   JWT_BLOCKING: a JSON Web Token (JWT) blacklist
-            /// *   IP_WHITELIST_CIDR: an IP address whitelist
+            /// *   JWT_BLOCKING : a JSON Web Token (JWT) blacklist
+            /// *   IP_WHITELIST_CIDR : an IP address whitelist
             /// *   PARAMETER_ACCESS: a list of parameters for parameter-based access control
             /// </summary>
             [NameInMap("DatasetType")]
@@ -52,14 +55,23 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
             [Validation(Required=false)]
             public string ModifiedTime { get; set; }
 
+            /// <summary>
+            /// The tags of the dataset.
+            /// </summary>
             [NameInMap("Tags")]
             [Validation(Required=false)]
             public List<DescribeDatasetListResponseBodyDatasetInfoListTags> Tags { get; set; }
             public class DescribeDatasetListResponseBodyDatasetInfoListTags : TeaModel {
+                /// <summary>
+                /// The tag key.
+                /// </summary>
                 [NameInMap("Key")]
                 [Validation(Required=false)]
                 public string Key { get; set; }
 
+                /// <summary>
+                /// The tag value.
+                /// </summary>
                 [NameInMap("Value")]
                 [Validation(Required=false)]
                 public string Value { get; set; }
@@ -68,18 +80,30 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 
         }
 
+        /// <summary>
+        /// The number of the page to return. Pages start from page 1. Default value: 1.
+        /// </summary>
         [NameInMap("PageNumber")]
         [Validation(Required=false)]
         public int? PageNumber { get; set; }
 
+        /// <summary>
+        /// The number of entries per page.
+        /// </summary>
         [NameInMap("PageSize")]
         [Validation(Required=false)]
         public int? PageSize { get; set; }
 
+        /// <summary>
+        /// The ID of the request.
+        /// </summary>
         [NameInMap("RequestId")]
         [Validation(Required=false)]
         public string RequestId { get; set; }
 
+        /// <summary>
+        /// The total number of entries returned.
+        /// </summary>
         [NameInMap("TotalCount")]
         [Validation(Required=false)]
         public int? TotalCount { get; set; }