+++ cloudapi-20160714/core/Models/ListTagResourcesRequest.cs
@@ -10,7 +10,7 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
 {
     public class ListTagResourcesRequest : TeaModel {
         /// <summary>
-        /// The token that determines the start point of the query.
+        /// The token used to start the next query.
         /// </summary>
         [NameInMap("NextToken")]
         [Validation(Required=false)]
@@ -19,7 +19,7 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
         /// <summary>
         /// The ID of the resource.
         /// 
-        /// Valid values of N: `[1, 50]`.
+        /// Valid values of N: `[1,50]`.
         /// </summary>
         [NameInMap("ResourceId")]
         [Validation(Required=false)]
@@ -28,7 +28,7 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
         /// <summary>
         /// The type of the resource.
         /// 
-        /// Tags are supported for API groups, plug-ins, and apps. Operations performed on APIs are based on the tags of API groups.
+        /// Currently, tags are supported for API groups, plug-ins, and apps. Operations performed on APIs are based on the tags of API groups.
         /// 
         /// *   **apiGroup**
         /// *   **plugin**
@@ -40,7 +40,7 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
         public string ResourceType { get; set; }
 
         /// <summary>
-        /// The tag of objects that match the rule. You can specify multiple tags.
+        /// The tag of objects that match the lifecycle rule. You can specify multiple tags.
         /// </summary>
         [NameInMap("Tag")]
         [Validation(Required=false)]
@@ -49,16 +49,16 @@ namespace AlibabaCloud.SDK.CloudAPI20160714.Models
             /// <summary>
             /// The key of the tag.
             /// 
-            /// Valid values of N: `[1, 20]`.
+            /// Valid values of N: `[1,20]`.
             /// </summary>
             [NameInMap("Key")]
             [Validation(Required=false)]
             public string Key { get; set; }
 
             /// <summary>
-            /// The value of the tag. If the parameter has a value, you must specify a value for the tag key with the same N as **tag.N.Key**. Otherwise, an error is reported.
+            /// The value of the tag. If the parameter has a value, a value is also required for the tag key of the same N in the corresponding **tag.N.Key**. Otherwise, an error is reported.
             /// 
-            /// Valid values of N: `[1, 20]`.
+            /// Valid values of N: `[1,20]`.
             /// </summary>
             [NameInMap("Value")]
             [Validation(Required=false)]